% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency.R
\name{currency}
\alias{currency}
\title{Convert a currency into another currency.}
\usage{
currency(amount, from = "USD", to = "USD")
}
\arguments{
\item{amount}{numeric. The amount to convert.}

\item{from}{character. The currency code to convert from. Defaults to \code{USD}.}

\item{to}{character. The currency code to convert to. Defaults to \code{USD}.}
}
\value{
the amount in the converted currency.
}
\description{
Convert a currency into another currency.
}

