% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency_codes.R
\name{get_code_from_unit}
\alias{get_code_from_unit}
\title{Get the currency code from a unit, using fuzzy matching.}
\usage{
get_code_from_unit(unit)
}
\arguments{
\item{unit}{character. The currency unit to try.}
}
\description{
Get the currency code from a unit, using fuzzy matching.
}
