% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency_codes.R
\name{get_unit_from_code}
\alias{get_unit_from_code}
\title{Get a unit from a currency code.}
\usage{
get_unit_from_code(code)
}
\arguments{
\item{code}{character. The currency code to use.}
}
\description{
Get a unit from a currency code.
}

