% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currency.R
\name{print.currency}
\alias{print.currency}
\title{Print for currency objects}
\usage{
\method{print}{currency}(x, ...)
}
\arguments{
\item{x}{currency. The currency object to print.}

\item{...}{list. Additional arguments to pass to print.}
}
\description{
Print for currency objects
}

